fix(compliance): register mandatory GDPR webhooks in the toml (App Store review blocker)
The 3 mandatory privacy webhooks (customers/data_request, customers/redact,
shop/redact) had handler files but were NOT declared anywhere — so they only
fired if set manually in the Partner Dashboard. Unregistered compliance
webhooks are an automatic App Store review failure.

- Declare them in shopify.app.toml as one [[webhooks.subscriptions]]
  compliance_topics block → /webhooks/compliance (Shopify's prescribed shape;
  shopify app deploy registers them).
- New consolidated webhooks.compliance.tsx: HMAC-verifies via authenticate.webhook
  (401 on bad signature — hard review requirement) and branches on topic
  (shop/redact deletes the merchant doc; customer topics acknowledge — the app
  stores no end-customer PII).
- Remove the 3 now-redundant single-topic handler files.
